Rename module to thread
