+++ homeassistant/util/thread.py
